added services catalogue filtering and completed admin functionality

+++ be/src/main/java/by/training/nc/sd2/reprository/ServiceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ServiceRepository extends CrudRepository<Service, Long> {
+public interface ServiceRepository extends CrudRepository<Service, Integer> {
 }
